requirements.txt
Flask~=3.1.2
flask_sqlalchemy~=3.1.1
python-dotenv~=1.1.1
